+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Kubernetes env provider can now read the trust bundle from both BinaryData and Data
+
+
 ## [0.3.2] - 2022-11-04
 ### Changed
 - Bugfix: Fix the kubebuilder enum annotations for NutanixTrustBundleKind